+++ changelog.md
@@ -1,1 +1,4 @@
-- Updated to 1.21.9
+- Updated required NeoForge version
+- Fixed breeder input items being sucked out with hoppers
+- Fixed converter input items being sucked out with hoppers
+- Fixed incubator input items being sucked out with hoppers